// src/app/core/enums/app-enums.ts
export enum RolesEnum {
  Custom = 0,
  Administrator = 1,
  Manager = 2,
  Staff = 3,
  Waiter = 4,
  DeliveryMan = 5,
}

export enum SectionEnum {
  Tables = 1,
  Menu = 2,
  Order = 3,
  Financials = 4,
  Reservation = 5,
  AddressBook = 6,
  Inventory = 7,
  Settings = 8,
  Delivery = 9,
  Kitchen = 10,
  Drawer = 11,
  Setup = 12,
}

export enum PermissionEnum {
  View = 1,
  Add = 2,
  Edit = 3,
  Delete = 4,
}